REF: collect arithmetic benchmarks (#32116)
