+++ asv_bench/benchmarks/categoricals.py
@@ -63,18 +63,6 @@ class Constructor:
         pd.Categorical(self.series)
 
 
-class CategoricalOps:
-    params = ["__lt__", "__le__", "__eq__", "__ne__", "__ge__", "__gt__"]
-    param_names = ["op"]
-
-    def setup(self, op):
-        N = 10 ** 5
-        self.cat = pd.Categorical(list("aabbcd") * N, ordered=True)
-
-    def time_categorical_op(self, op):
-        getattr(self.cat, op)("b")
-
-
 class Concat:
     def setup(self):
         N = 10 ** 5
