+++ asv_bench/benchmarks/timedelta.py
@@ -5,7 +5,7 @@ benchmarks.tslibs.timedelta for benchmarks that rely only on tslibs.
 
 import numpy as np
 
-from pandas import DataFrame, Series, Timestamp, timedelta_range, to_timedelta
+from pandas import DataFrame, Series, timedelta_range, to_timedelta
 
 
 class ToTimedelta:
@@ -39,15 +39,6 @@ class ToTimedeltaErrors:
 
     def time_convert(self, errors):
         to_timedelta(self.arr, errors=errors)
-
-
-class TimedeltaOps:
-    def setup(self):
-        self.td = to_timedelta(np.arange(1000000))
-        self.ts = Timestamp("2000")
-
-    def time_add_td_ts(self):
-        self.td + self.ts
 
 
 class DatetimeAccessor:
